feature(repository): edit query statistics

+++ hrm/src/main/java/com/project/hrm/dto/statisticsDTO/TotalContractByStatus.java
@@ -2,5 +2,5 @@ package com.project.hrm.dto.statisticsDTO;
 
 public interface TotalContractByStatus {
     String getStatus();
-    Integer getTotal();
+    Long getTotal();
 }
